+++ src/main/java/blackjack/domain/gameplayer/Dealer.java
@@ -1,6 +1,5 @@
 package blackjack.domain.gameplayer;
 
-import blackjack.domain.Score;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
