+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
+import blackjack.domain.Game;
 import blackjack.domain.gameplayer.Dealer;
 import blackjack.domain.gameplayer.Player;
