+++ src/test/java/blackjack/domain/gameplayer/ScoreTest.java
@@ -1,9 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.gameplayer;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardSymbol;
 import blackjack.domain.card.Cards;
+import blackjack.domain.gameplayer.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
